refactor(test): relocate shared KV pool config
Keep nightly-only KV pool configuration under a dedicated package shared by single-node and multi-node test frameworks.
