+++ tests/e2e/nightly/kv_pool/__init__.py
@@ -1,0 +1,1 @@
+"""Shared KV pool support for nightly E2E test frameworks."""
