+++ tests/e2e/nightly/single_node/models/scripts/single_node_config.py
@@ -7,7 +7,7 @@ import regex as re
 import yaml
 from vllm.utils.network_utils import get_open_port
 
-from tests.e2e.kv_pool import KVPoolConfig, parse_kv_pool_config
+from tests.e2e.nightly.kv_pool.config import KVPoolConfig, parse_kv_pool_config
 
 CONFIG_BASE_PATH = os.getenv("CONFIG_BASE_PATH") or "tests/e2e/nightly/single_node/models/configs"
